[onert] Separate IOperation from the Operation header
This commit separates IOperation from the Operation header.

ONE-DCO-1.0-Signed-off-by: ragmani <ragmani0216@gmail.com>
